third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/contentSecurityPolicy/redirect-with-delay-expected.txt

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/script-redirect-not-allowed.js' because it violates the following Content Security Policy directive: "script-src http://localhost:8000/ 'unsafe-inline'".

This is a testharness.js-based test.
PASS CSP should be enforced to the iframe 
Harness: the test ran to completion.
